==> notebooks/00_exploration.ipynb <==
DATA_PATH = "../data/raw/survey_results_public.csv"

import pandas as pd
import numpy as np
pd.options.display.max_rows = 1000

# Read data and print shape
raw_df = pd.read_csv(DATA_PATH)
raw_df.shape

#display random answer to get an intution about the data
#Observations: Multiple answers need to be splitted
raw_df.sample(1).iloc[0]

# Print the general info. of the data frame
raw_df.info()

**Observe that most coulumns are formatted as objects; we need to deal with that**

#Get stats of the numerical columns
raw_df.describe()

# Lets take a look at the questionable columns
questionable_cols = ['YearsCodePro', 'YearsCode']

for col in questionable_cols:
    print(col)
    print(raw_df[col].unique().tolist())
    print('------------------------------\n')    

**We can notice that the columns has some values which need to be changed like 'More than 50 years' we can change into 51 or something like that**
