==> notebooks/01_preprocessing.ipynb <==
# Constants
DATA_PATH = "../data/raw/survey_results_public.csv"
EXPORT_PATH = "../data/processed/1_preprocessed_df.pkl"

REPLACE_DICT = {
    'YearsCodePro': {'Less than 1 year': 0, 'More than 50 years': 51},
    'YearsCode':    {'Less than 1 year': 0, 'More than 50 years': 51}}

# Needed packages
import pandas as pd
import numpy as np
import logging
import pickle

---
## Functions

def split_answers(data_series, delimiter=';'):
    """
    Split multiple answers exists in a single string into a list
    of strings each represinting a single answer
    
    Parameters:
    * data_series (pd.series): String series with answers
    * delimeter (string): Another decimal integer, defalut to (;)
    
    Returns:
    * (pd.series) if column contains multiple answers
    """
    
    #Sub functions
    def is_splittable(pd_series, delimiter):
        """check if column has multiple answers that could be splitted or not"""
        return pd_series.str.contains(delimiter)
    
    def split_answer(pd_series, delimiter):
        """Function to split single answer"""
        return pd_series.str.split(delimiter)
    
    # --------------------------------------
    
    # Check if multiple answers esist - if not; return original
    splittable_values = is_splittable(data_series, delimiter)
    if not splittable_values.any():
        return data_series
    
    #Else, split each value to a list
    modified_series = split_answer(data_series, delimiter)
    
    #Replace NAs with empty lists
    mask_null = modified_series.isnull()
    modified_series.loc[mask_null] = modified_series.loc[mask_null].apply(lambda x: [])
    
    return modified_series

---
## Processing

# reading data
raw_df = pd.read_csv(DATA_PATH)
df = raw_df.copy()

### 1. Replace values and parse

# Replace the non-numerical values in colmuns into numbers
for col, replacement in REPLACE_DICT.items():
    df[col] = df[col].replace(replacement).astype(np.float32)

### 2. Split multiple answers

object_cols = df.select_dtypes(include="object").columns.tolist()
for col in object_cols:
    df[col] = split_answers(df[col])

---
## Visually verify results

i = df.sample(1).index[0]
print(raw_df['LanguageHaveWorkedWith'].iloc[i])
print(df['LanguageHaveWorkedWith'].iloc[i])

i = df.sample(1).index[0]
print(raw_df['DevType'].iloc[i])
print(df['DevType'].iloc[i])

i = df.sample(1).index[0]
print(raw_df['YearsCodePro'].iloc[i])
print(df['YearsCodePro'].iloc[i])

---
# Export Data

df.to_pickle(EXPORT_PATH)
