==> notebooks/04c_model_job_rf.ipynb <==
DF_PATH = "../data/processed/4_balanced_data.pkl"

ROLE_COLS      = ['DevType']
TECH_COLS      = ['LanguageHaveWorkedWith',
                  'DatabaseHaveWorkedWith',
                  'PlatformHaveWorkedWith',
                  'WebframeHaveWorkedWith',
                  'MiscTechHaveWorkedWith',
                  'ToolsTechHaveWorkedWith',
                  'NEWCollabToolsHaveWorkedWith']

MLFLOW_TRACKING_URI = '../models/mlruns'
MLFLOW_EXPERIMENT_NAME = "skills_jobs_stackoverflow"

LOG_PATH = "../models/temp/"
LOG_DATA_PKL = "data.pkl"
LOG_MODEL_PKL = "model.pkl"
LOG_METRICS_PKL = "metrics.pkl"

import pandas as pd
import numpy as np
import os
import pickle

import mlflow
from mlflow.tracking import MlflowClient

from sklearn.model_selection import train_test_split
from sklearn.model_selection import GridSearchCV
from sklearn.preprocessing import RobustScaler
from sklearn.decomposition import PCA
from sklearn.pipeline import make_pipeline
from sklearn.metrics import accuracy_score, precision_score, recall_score, f1_score

from sklearn.ensemble import RandomForestClassifier

pd.options.display.max_columns=100

___
# Functions

def calculate_quality(ground_truth, predictoins, metric_function, sort_values=False):
    """
    Calculate the quality of your model according to metric score
    instead of giving a summary of overall quality
    Input:
        ground_truth: from real data
        predictions: the predicted from the model
        metric_function: the metric score function used to measure performance
    Output:
        a dict of all scores w.r.t every class (column) separately
    """
    quality_scores = {}
    for col in predictions.columns:
        role_truth = ground_truth[col].copy()
        role_pred = predictions[col].copy()
        quality_scores[col] = round(metric_function(role_truth, role_pred) * 100, 2)
    
    quality_scores = pd.Series(quality_scores.values(), index=quality_scores.keys())
    if sort_values:
        quality_scores = quality_scores.sort_values()
    
    return quality_scores

___
# Prepare Data:
## Read data:

# Read data
df = pd.read_pickle(DF_PATH)

## Split Data to train and test:

X_train, X_test, y_train, y_test = train_test_split(df.drop(['DevType'], axis=1),
                                                    df['DevType'],
                                                    random_state=0)

---
# Train models
## Intialize MLflow

#Initialize client and experiment
mlflow.set_tracking_uri(MLFLOW_TRACKING_URI)
client = MlflowClient()

#Check if experiment already exists; if not create one
try:
    mlflow.create_experiment(MLFLOW_EXPERIMENT_NAME)
except:
    print(f'Experiment "{MLFLOW_EXPERIMENT_NAME}" exists at "{mlflow.get_tracking_uri()}"')

exp = client.get_experiment_by_name(MLFLOW_EXPERIMENT_NAME)
#print(mlflow.get_experiment_by_name(MLFLOW_EXPERIMENT_NAME))
print(exp)

## 1. Vanilla Forest

rf_clf = make_pipeline(RobustScaler(),
                       PCA(n_components=0.95),
                       RandomForestClassifier(n_jobs=8,
                                              verbose=1,
                                              random_state=0))

rf_clf.fit(X_train.values, y_train.values)

# Evaluate on training set
predictions = pd.DataFrame(rf_clf.predict(X_train.values),
                           columns = y_train.columns)

train_scores = {score.__name__: calculate_quality(y_train, predictions, score)
                for score in [accuracy_score, precision_score, recall_score, f1_score]}
train_scores = pd.concat(train_scores, axis=1)
mean_train_score = train_scores.mean()

print(mean_train_score)
train_scores

# Evaluate on test set
predictions = pd.DataFrame(rf_clf.predict(X_test.values), 
                                       columns=y_test.columns)

test_scores = {score.__name__: calculate_quality(y_test, predictions, score)
               for score in [accuracy_score, precision_score, recall_score, f1_score]}
test_scores = pd.concat(test_scores, axis=1)
mean_test_scores = test_scores.mean()

print(mean_test_scores)
test_scores


## Prepare Logs

# Data details
data_details = {"data_path": DF_PATH,
                "training_indices": X_train.index.tolist(),
                "test_indices": X_test.index.tolist(),
                "features_names": X_train.columns.droplevel(0).tolist(),
                "targets_names": y_train.columns.tolist()}

with open(os.path.join(LOG_PATH, LOG_DATA_PKL), 'wb') as output_file:
    pickle.dump(data_details, output_file)

# Model
model = {"model_description": "Random Forest: with PCA - Basic",
         "model_details": str(rf_clf),
         "model_object": rf_clf}

with open(os.path.join(LOG_PATH, LOG_MODEL_PKL), 'wb') as output_file:
    pickle.dump(model, output_file)

# Performance details
classes_metrics = {"train_scores": train_scores,
                   "test_scores": test_scores}

with open(os.path.join(LOG_PATH, LOG_METRICS_PKL), 'wb') as output_file:
    pickle.dump(classes_metrics, output_file)

## Logging

# Start a new run and track
with mlflow.start_run(experiment_id=exp.experiment_id,
                      run_name=model['model_description']):
    #Log pickles
    mlflow.log_artifacts(LOG_PATH)
    
    #Track metrics
    for metric, score in mean_test_scores.items():
        mlflow.log_metric(metric, score)

___
# Hyper parameter tuning:
**Here we use grid search to tune our parameters**

hpt_rf_clf = make_pipeline(RobustScaler(),
                           PCA(),
                           RandomForestClassifier(n_jobs=8,
                                                  verbose=1,
                                                  random_state=0))

# Get all the model patameter to specify which to tune
list(hpt_rf_clf.get_params().keys())

# Write down the values to tune on
tuned_parameters = [{
    'pca__n_components': [0.7, 0.85, 0.95],
    'randomforestclassifier__n_estimators': [250, 500],
    'randomforestclassifier__max_depth': [3, 10, None],
}]

hpt_rf_clf = GridSearchCV(hpt_rf_clf, tuned_parameters)
hpt_rf_clf.fit(X_train.values, y_train.values)

# Print the tuned model parameters
hpt_rf_clf.best_params_

## Evaluate the model

# Evaluate on training set
predictions = pd.DataFrame(hpt_rf_clf.predict(X_train.values),
                           columns = y_train.columns)

train_scores = {score.__name__: calculate_quality(y_train, predictions, score)
                for score in [accuracy_score, precision_score, recall_score, f1_score]}
train_scores = pd.concat(train_scores, axis=1)
mean_train_scores = train_scores.mean()

print(mean_train_scores)
train_scores

# Evaluate on test set
predictions = pd.DataFrame(hpt_rf_clf.predict(X_test.values), 
                                       columns=y_test.columns)

test_scores = {score.__name__: calculate_quality(y_test, predictions, score)
               for score in [accuracy_score, precision_score, recall_score, f1_score]}
test_scores = pd.concat(test_scores, axis=1)
mean_test_scores = test_scores.mean()

print(mean_test_scores)
test_scores

## Prepare logs:

# Data details
data_details = {"data_path": DF_PATH,
                "training_indices": X_train.index.tolist(),
                "test_indices":     X_test.index.tolist(), 
                "features_names":   X_train.columns.droplevel(0).tolist(),
                "targets_names":    y_train.columns.tolist()}

with open(os.path.join(LOG_PATH, LOG_DATA_PKL), "wb") as output_file:
    pickle.dump(data_details, output_file)

# Model
model = {"model_description": "Random Forest: with PCA + Hyperparamter tuning",
         "model_details": str(hpt_rf_clf),
         "model_object": hpt_rf_clf} 

with open(os.path.join(LOG_PATH, LOG_MODEL_PKL), "wb") as output_file:
    pickle.dump(model, output_file)

# Preformance details
classes_metrics = {"train_scores": train_scores, 
                   "test_scores":  test_scores}

with open(os.path.join(LOG_PATH, LOG_METRICS_PKL), "wb") as output_file:
    pickle.dump(classes_metrics, output_file)

## Logginig:

# Start a new run and track 
with mlflow.start_run(experiment_id=exp.experiment_id, 
                      run_name=model["model_description"]):
    # Log pickles
    mlflow.log_artifacts(LOG_PATH)  
    
    # Track metrics 
    for metric, score in mean_test_scores.items():
        mlflow.log_metric(metric, score)
