==> notebooks/05a_prediction_notebook.ipynb <==
MLFLOW_TRACKING_URI = "../models/mlruns"
MLFLOW_RUN_ID = "a8fb8540585f471ab7bf110e5a28229a"

LOG_DATA_PKL = "data.pkl"
LOG_MODEL_PKL = "model.pkl"
LOG_METRIC_PKL = "metrics.pkl"

CLUSTERS_YAML_PATH = "../data/processed/features_skills_clusters_description.yaml"

import pandas as pd

import os 
import pickle
import yaml

import mlflow
from mlflow.tracking import MlflowClient

---
# Intilaize:
## Mlflow:

# Intialize client
mlflow.set_tracking_uri(MLFLOW_TRACKING_URI)
client = MlflowClient()

run = mlflow.get_run(MLFLOW_RUN_ID)
artifacts_path = run.info.artifact_uri

artifacts_path

### Load artifacts

# Load model
model_path = os.path.join(artifacts_path, LOG_MODEL_PKL)
with open(model_path, 'rb') as file:
    model = pickle.load(file)

model.keys()

# Load data pickle
data_path = os.path.join(artifacts_path, LOG_DATA_PKL)
with open(data_path, 'rb') as file:
    data = pickle.load(file)

data.keys()

# Unpack variables:
features_names = pd.Series(data["features_names"])
targets_names = pd.Series(data["targets_names"])
classifier = model["model_object"]

---
## Load skills Clusters

# Load skills clusters
with open(CLUSTERS_YAML_PATH, 'r') as stream:
    cluster_config = yaml.safe_load(stream)

print([print(cluster, skills) for cluster, skills in cluster_config.items()])

# Reformat into dataframe:
molten_clusters = [(cluster_name, cluster_skill)
                   for cluster_name, cluster_skills in cluster_config.items()
                   for cluster_skill in cluster_skills]

clusters_df = pd.DataFrame(molten_clusters, columns=["cluster_name", "skill"])
clusters_df

## Predict sample entry

sample_skills = ["Unreal Engine", "Unity 3D"]#, "Hadoop", "Python"]

# Verify that these skills existis as features
pd.Series(sample_skills).isin(features_names)

### 1. Recreate cluster features

sample_clusters = clusters_df.copy()
sample_clusters["sample_skills"] = sample_clusters["skill"].isin(sample_skills)
sample_clusters

cluster_features = sample_clusters.groupby("cluster_name")["sample_skills"].sum()
cluster_features

### 2. Create OneHotEncoded skills:

# First take only the skills columns; we drop only cluster_features from the overall features_names
skills_names = features_names[~features_names.isin(cluster_features.index)]
skills_names

# Now using the skills_names we can construct a pandas series;
# that contains all skills with 1 for skills in the sample skills
ohe_skills = pd.Series(skills_names.isin(sample_skills).astype(int).tolist(),
                       index=skills_names)
ohe_skills

### 3. Combine features

# Concat 
features = pd.concat([ohe_skills, cluster_features])

# Sort columns according to features_names
features = features.loc[features_names]
features

### 4. Predict

predictions = classifier.predict_proba([features.values])
predictions

# Now print the props sorted and indexed with target names
positive_probs = [prob[0][1] for prob in predictions]
pd.Series(positive_probs,
          index=targets_names).sort_values(ascending=False)
