==> notebooks/05b_simulation_notebook.ipynb <==
MLFLOW_TRACKING_URI = "../models/mlruns"
MLFLOW_RUN_ID = "a8fb8540585f471ab7bf110e5a28229a"

CLUSTERS_YAML_PATH = "../data/processed/features_skills_clusters_description.yaml"

import pandas as pd
import sys

sys.path.append("../scripts/")

import warnings
warnings.filterwarnings("ignore")
#from JobDeena import *
from JobPrediction import *

---
# Simulate additional skills
## 1. Calculate base probs

# Intialize model
model = JobPrediction(MLFLOW_TRACKING_URI, MLFLOW_RUN_ID, CLUSTERS_YAML_PATH)

available_skills = ["Unreal Engine", "Unity 3D"]#["Hadoop", "Python"]#,  "Git"]

base_predictions = model.predict_jobs_probabilities(available_skills)
base_predictions.sort_values(ascending=False)

## 2. Calculate Probs for new skills:

# Get possible new skills
all_skills = pd.Series(model.get_all_skills())
new_skills = all_skills[~all_skills.isin(available_skills)].copy()

new_skills

simulated_results = []
for skill in new_skills:
    additional_skills_prob = model.predict_jobs_probabilities([skill] + available_skills)
    additional_skill_uplift = (additional_skills_prob - base_predictions) / base_predictions
    additional_skill_uplift.name = skill
    simulated_results.append(additional_skill_uplift)

simulated_results = pd.DataFrame(simulated_results)
simulated_results

## Recommend

#target_job = "Data scientist or machine learning specialist"
target_job = "Engineer, data"

target_results = simulated_results[target_job].sort_values(ascending=False)
target_results

threshold = 0
recommendations = target_results[target_results > threshold].index.tolist()

recommendations

print(f"Your current skills: {str(available_skills)}")
print(f"Your target job: {str(target_job)}")
print(f"You might also consider learning: {recommendations}")
