==> notebooks/06_api_test.ipynb <==
API_URL = "http://127.0.0.1:5000/"
PREDICTION_ENDPOINT = "predict_jobs_probs"
SIMULATION_ENDPOINT = "recommend_new_skills"

import pandas as pd 
import json
import requests

## Predict

available_skills = ["R", "SQL", "Rstudio"]

predictions = requests.post(API_URL + PREDICTION_ENDPOINT,
                            data=json.dumps(available_skills),
                            headers={"content-type": "application/json"})

predictions

pd.Series(predictions.json()).sort_values(ascending=False)

## Simulate

target_job = "Data scientist or machine learning specialist"

simulation = requests.post(API_URL + SIMULATION_ENDPOINT,
                           data = json.dumps({
                               "available_skills": available_skills,
                               "target_job": target_job
                           }),
                           headers={"content-type": "application/json"})

pd.Series(simulation.json()).sort_values(ascending=False)[:10]
